--- app/views/layouts/main.php
@@ -1,15 +1,0 @@
-<!DOCTYPE html>
-<?php 
-/* @var $this DView */
-/* @var $content string */
-?>
-<html>
-	<head>
-		<title><?php echo $this->title; ?></title>
-		<meta charset="UTF-8">
-		<meta name="viewport" content="width=device-width, initial-scale=1.0">
-	</head>
-	<body>
-        <div><?php echo $content;?></div>
-	</body>
-</html>
