+++ apps/basic/protected/config/main.php
@@ -1,0 +1,13 @@
+<?php
+
+return [
+    'basePath' => dirname(__DIR__),
+    'defaultRoute' => 'home',
+    'components' => [
+        'db' => [
+            'dsn' => 'mysql:host=localhost;dbname=mydb',
+            'username' => '',
+            'password' => ''
+        ]
+    ]
+];
