+++ apps/basic/protected/controllers/PostController.php
@@ -1,0 +1,25 @@
+<?php
+
+namespace app\controllers;
+
+use dee\core\Controller;
+
+/**
+ * Description of PostController
+ *
+ * @author Misbahul D Munir <misbahuldmunir@gmail.com>
+ * @since 1.0
+ */
+class PostController extends Controller
+{
+
+    public function actionIndex()
+    {
+
+    }
+
+    public function actionView($id)
+    {
+        
+    }
+}
