+++ apps/basic/protected/views/layouts/main.php
@@ -1,0 +1,28 @@
+<!DOCTYPE html>
+<?php 
+/* @var $this dee\base\View */
+/* @var $content string */
+?>
+<html>
+    <head>
+        <title><?= $this->title ?></title>
+        <meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
+        <link href="//maxcdn.bootstrapcdn.com/bootstrap/3.2.0/css/bootstrap.min.css" rel="stylesheet">
+        <script type="text/javascript" src="//code.jquery.com/jquery-2.1.1.min.js"></script>
+    </head>
+    <body>
+        <div class="wrap">
+            <div class="container">
+                <div class="pull-right">
+                    <?php if (isset($_SESSION['__identity'])): ?>
+                        <?= $_SESSION['__identity'] ?> &nbsp; <a href="<?= createUrl('logout') ?>">Logout</a>
+                    <?php else: ?>
+                        <a href="<?= createUrl('login') ?>">Login</a>
+                    <?php endif; ?>
+                </div>
+                <h1><?= $this->title ?></h1>
+                <?= $content ?>
+            </div>
+        </div>
+    </body>
+</html>
