+++ apps/benchmark/protected/.htaccess
@@ -1,0 +1,1 @@
+deny from all
