+++ apps/benchmark/protected/controllers/SiteController.php
@@ -1,0 +1,15 @@
+<?php
+
+namespace app\controllers;
+
+use dee\web\Controller;
+
+class SiteController extends Controller
+{
+    public $defaultAction = 'hello';
+
+    public function actionHello()
+    {
+        return 'hello world';
+    }
+}
