+++ framework/Dee.php
@@ -11,5 +11,6 @@ class Dee extends \dee\BaseDee
 {
     
 }
+
 Dee::$classMap = require(__DIR__ . '/classes.php');
 spl_autoload_register(['Dee', 'autoload']);
