--- framework/core/Configurable.php
@@ -1,18 +1,0 @@
-<?php
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
-
-namespace dee\core;
-
-/**
- *
- * @author Misbahul D Munir <misbahuldmunir@gmail.com>
- * @since 1.0
- */
-interface Configurable
-{
-    
-}
