--- framework/core/ErrorHandler.php
@@ -1,245 +1,0 @@
-<?php
-
-namespace dee\core;
-
-use ErrorException;
-use Exception;
-
-/**
- * Description of ErrorHandler
- *
- * @author Misbahul D Munir <misbahuldmunir@gmail.com>
- * @since 1.0
- */
-class ErrorHandler extends Object
-{
-    /**
-     * @var boolean whether to discard any existing page output before error display. Defaults to true.
-     */
-    public $discardExistingOutput = true;
-
-    /**
-     * @var integer the size of the reserved memory. A portion of memory is pre-allocated so that
-     * when an out-of-memory issue occurs, the error handler is able to handle the error with
-     * the help of this reserved memory. If you set this value to be 0, no memory will be reserved.
-     * Defaults to 256KB.
-     */
-    public $memoryReserveSize = 262144;
-
-    /**
-     * @var \Exception the exception that is being handled currently.
-     */
-    public $exception;
-
-    /**
-     * @var string Used to reserve memory for fatal error handler.
-     */
-    private $_memoryReserve;
-
-    /**
-     * Register this error handler
-     */
-    public function register()
-    {
-        ini_set('display_errors', false);
-        set_exception_handler([$this, 'handleException']);
-        set_error_handler([$this, 'handleError']);
-        if ($this->memoryReserveSize > 0) {
-            $this->_memoryReserve = str_repeat('x', $this->memoryReserveSize);
-        }
-        register_shutdown_function([$this, 'handleFatalError']);
-    }
-
-    /**
-     * Unregisters this error handler by restoring the PHP error and exception handlers.
-     */
-    public function unregister()
-    {
-        restore_error_handler();
-        restore_exception_handler();
-    }
-
-    /**
-     * Handles uncaught PHP exceptions.
-     *
-     * This method is implemented as a PHP exception handler.
-     *
-     * @param \Exception $exception the exception that is not caught
-     */
-    public function handleException($exception)
-    {
-        $this->exception = $exception;
-
-        // disable error capturing to avoid recursive errors while handling exceptions
-        $this->unregister();
-
-        try {
-            $this->logException($exception);
-            if ($this->discardExistingOutput) {
-                $this->clearOutput();
-            }
-            $this->renderException($exception);
-            if (!DEE_ENV_TEST) {
-                exit(1);
-            }
-        } catch (Exception $e) {
-            // an other exception could be thrown while displaying the exception
-            $msg = (string) $e;
-            $msg .= "\nPrevious exception:\n";
-            $msg .= (string) $exception;
-            if (DEE_DEBUG) {
-                if (PHP_SAPI === 'cli') {
-                    echo $msg . "\n";
-                } else {
-                    echo '<pre>' . htmlspecialchars($msg, ENT_QUOTES, 'UTF-8') . '</pre>';
-                }
-            } else {
-                echo 'An internal server error occurred.';
-            }
-            $msg .= "\n\$_SERVER = " . var_export($_SERVER, true);
-            error_log($msg);
-
-            if (PHP_SAPI !== 'cli') {
-                http_response_code(500);
-            }
-            exit(1);
-        }
-
-        $this->exception = null;
-    }
-
-    /**
-     * Handles PHP execution errors such as warnings and notices.
-     *
-     * This method is used as a PHP error handler. It will simply raise an [[ErrorException]].
-     *
-     * @param integer $code the level of the error raised.
-     * @param string $message the error message.
-     * @param string $file the filename that the error was raised in.
-     * @param integer $line the line number the error was raised at.
-     * @return boolean whether the normal error handler continues.
-     *
-     * @throws ErrorException
-     */
-    public function handleError($code, $message, $file, $line)
-    {
-        if (error_reporting() & $code) {
-            // load ErrorException manually here because autoloading them will not work
-            // when error occurs while autoloading a class
-            $exception = new ErrorException($message, $code, $code, $file, $line);
-
-            // in case error appeared in __toString method we can't throw any exception
-            $trace = debug_backtrace(DEBUG_BACKTRACE_IGNORE_ARGS);
-            array_shift($trace);
-            foreach ($trace as $frame) {
-                if ($frame['function'] == '__toString') {
-                    $this->handleException($exception);
-                    exit(1);
-                }
-            }
-
-            throw $exception;
-        }
-        return false;
-    }
-
-    /**
-     * Handles fatal PHP errors
-     */
-    public function handleFatalError()
-    {
-        unset($this->_memoryReserve);
-
-        $error = error_get_last();
-
-        $fatalError = isset($error['type']) && in_array($error['type'], [E_ERROR, E_PARSE, E_CORE_ERROR, E_CORE_WARNING,
-                E_COMPILE_ERROR, E_COMPILE_WARNING]);
-        if ($fatalError) {
-            $exception = new ErrorException($error['message'], $error['type'], $error['type'], $error['file'], $error['line']);
-            $this->exception = $exception;
-
-            $this->logException($exception);
-
-            if ($this->discardExistingOutput) {
-                $this->clearOutput();
-            }
-            $this->renderException($exception);
-
-            exit(1);
-        }
-    }
-
-    /**
-     * Renders the exception.
-     * @param \Exception $exception the exception to be rendered.
-     */
-    protected function renderException($exception)
-    {
-        $trace = $this->getTrace($exception);
-        echo implode('<br>', $trace);
-    }
-
-    /**
-     * Renders the exception.
-     * @param \Exception $exception the exception to be rendered.
-     */
-    protected function getTrace($exception)
-    {
-        $result = [];
-        $result[] = 'Error: ' . $exception->getMessage();
-        $result[] = '';
-        for ($i = 0, $trace = $exception->getTrace(), $length = count($trace); $i < $length; ++$i) {
-            if (isset($trace[$i]['file'], $trace[$i]['line'])) {
-                $result[] = $trace[$i]['file'] . ' at line ' . $trace[$i]['line'];
-            }
-        }
-        return $result;
-    }
-
-    /**
-     * Logs the given exception
-     * @param \Exception $exception the exception to be logged
-     * @since 2.0.3 this method is now public.
-     */
-    public function logException($exception)
-    {
-        
-    }
-
-    /**
-     * Removes all output echoed before calling this method.
-     */
-    public function clearOutput()
-    {
-        // the following manual level counting is to deal with zlib.output_compression set to On
-        for ($level = ob_get_level(); $level > 0; --$level) {
-            if (!@ob_end_clean()) {
-                ob_clean();
-            }
-        }
-    }
-
-    /**
-     * Converts an exception into a PHP error.
-     *
-     * This method can be used to convert exceptions inside of methods like `__toString()`
-     * to PHP errors because exceptions cannot be thrown inside of them.
-     * @param \Exception $exception the exception to convert to a PHP error.
-     */
-    public static function convertExceptionToError($exception)
-    {
-        trigger_error(static::convertExceptionToString($exception), E_USER_ERROR);
-    }
-
-    /**
-     * Converts an exception into a simple string.
-     * @param \Exception $exception the exception being converted
-     * @return string the string representation of the exception.
-     */
-    public static function convertExceptionToString($exception)
-    {
-        $message = 'Error: ' . $exception->getMessage();
-
-        return $message;
-    }
-}
