--- framework/core/Object.php
@@ -1,167 +1,0 @@
-<?php
-
-namespace dee\core;
-
-use Dee;
-use Exception;
-
-/**
- * Description of Object
- *
- * @author Misbahul D Munir <misbahuldmunir@gmail.com>
- * @since 1.0
- */
-class Object implements Configurable
-{
-
-    public function __construct($config = [])
-    {
-        Dee::configure($this, $config);
-        $this->init();
-    }
-
-    public function init()
-    {
-
-    }
-
-    /**
-     * Returns the value of an object property.
-     *
-     * Do not call this method directly as it is a PHP magic method that
-     * will be implicitly called when executing `$value = $object->property;`.
-     * @param  string    $name the property name
-     * @return mixed     the property value
-     * @throws Exception if the property is not defined or write-only
-     * @see __set()
-     */
-    public function __get($name)
-    {
-        $getter = 'get' . $name;
-        if (method_exists($this, $getter)) {
-            return $this->$getter();
-        } elseif (method_exists($this, 'set' . $name)) {
-            throw new Exception('Getting write-only property: ' . get_class($this) . '::' . $name);
-        } else {
-            throw new Exception('Getting unknown property: ' . get_class($this) . '::' . $name);
-        }
-    }
-
-    /**
-     * Sets value of an object property.
-     *
-     * Do not call this method directly as it is a PHP magic method that
-     * will be implicitly called when executing `$object->property = $value;`.
-     * @param  string    $name  the property name or the event name
-     * @param  mixed     $value the property value
-     * @throws Exception if the property is not defined or is read-only
-     * @see __get()
-     */
-    public function __set($name, $value)
-    {
-        $setter = 'set' . $name;
-        if (method_exists($this, $setter)) {
-            $this->$setter($value);
-        } elseif (method_exists($this, 'get' . $name)) {
-            throw new Exception('Setting read-only property: ' . get_class($this) . '::' . $name);
-        } else {
-            throw new Exception('Setting unknown property: ' . get_class($this) . '::' . $name);
-        }
-    }
-
-    /**
-     * Checks if the named property is set (not null).
-     *
-     * Do not call this method directly as it is a PHP magic method that
-     * will be implicitly called when executing `isset($object->property)`.
-     *
-     * Note that if the property is not defined, false will be returned.
-     * @param  string  $name the property name or the event name
-     * @return boolean whether the named property is set (not null).
-     */
-    public function __isset($name)
-    {
-        $getter = 'get' . $name;
-        if (method_exists($this, $getter)) {
-            return $this->$getter() !== null;
-        } else {
-            return false;
-        }
-    }
-
-    /**
-     * Sets an object property to null.
-     *
-     * Do not call this method directly as it is a PHP magic method that
-     * will be implicitly called when executing `unset($object->property)`.
-     *
-     * Note that if the property is not defined, this method will do nothing.
-     * If the property is read-only, it will throw an exception.
-     * @param  string    $name the property name
-     * @throws Exception if the property is read only.
-     */
-    public function __unset($name)
-    {
-        $setter = 'set' . $name;
-        if (method_exists($this, $setter)) {
-            $this->$setter(null);
-        } elseif (method_exists($this, 'get' . $name)) {
-            throw new Exception('Unsetting read-only property: ' . get_class($this) . '::' . $name);
-        }
-    }
-
-    /**
-     * Returns a value indicating whether a property is defined.
-     * A property is defined if:
-     *
-     * - the class has a getter or setter method associated with the specified name
-     *   (in this case, property name is case-insensitive);
-     * - the class has a member variable with the specified name (when `$checkVars` is true);
-     *
-     * @param  string  $name      the property name
-     * @param  boolean $checkVars whether to treat member variables as properties
-     * @return boolean whether the property is defined
-     * @see canGetProperty()
-     * @see canSetProperty()
-     */
-    public function hasProperty($name, $checkVars = true)
-    {
-        return $this->canGetProperty($name, $checkVars) || $this->canSetProperty($name, false);
-    }
-
-    /**
-     * Returns a value indicating whether a property can be read.
-     * A property is readable if:
-     *
-     * - the class has a getter method associated with the specified name
-     *   (in this case, property name is case-insensitive);
-     * - the class has a member variable with the specified name (when `$checkVars` is true);
-     *
-     * @param  string  $name      the property name
-     * @param  boolean $checkVars whether to treat member variables as properties
-     * @return boolean whether the property can be read
-     * @see canSetProperty()
-     */
-    public function canGetProperty($name, $checkVars = true)
-    {
-        return method_exists($this, 'get' . $name) || $checkVars && property_exists($this, $name);
-    }
-
-    /**
-     * Returns a value indicating whether a property can be set.
-     * A property is writable if:
-     *
-     * - the class has a setter method associated with the specified name
-     *   (in this case, property name is case-insensitive);
-     * - the class has a member variable with the specified name (when `$checkVars` is true);
-     *
-     * @param  string  $name      the property name
-     * @param  boolean $checkVars whether to treat member variables as properties
-     * @return boolean whether the property can be written
-     * @see canGetProperty()
-     */
-    public function canSetProperty($name, $checkVars = true)
-    {
-        return method_exists($this, 'set' . $name) || $checkVars && property_exists($this, $name);
-    }
-}
