+++ framework/web/Request.php
@@ -1,0 +1,426 @@
+<?php
+
+namespace dee\web;
+
+/**
+ * Description of Request
+ *
+ * @author Misbahul D Munir <misbahuldmunir@gmail.com>
+ * @since 1.0
+ */
+class Request extends \dee\base\Request
+{
+    public $enablePrettyUrl = false;
+    public $cookieValidationKey;
+    public $routeParam = 'r';
+    public $parsers = [];
+
+    public function init()
+    {
+        if ($this->cookieValidationKey === null) {
+            $this->cookieValidationKey = Helper::getRandomKey('request.validationKey');
+        }
+    }
+
+    public function resolve()
+    {
+        if ($this->enablePrettyUrl) {
+            return [$this->getPathInfo(), $this->get()];
+        } else {
+            return [$this->get($this->routeParam, ''), $this->get()];
+        }
+    }
+    private $_bodyParams;
+
+    public function post($name = null, $default = null)
+    {
+        if ($this->_bodyParams === null) {
+            $this->_bodyParams = [];
+            $rawBody = file_get_contents('php://input');
+            $contentType = $this->getContentType();
+            if ($contentType && ($pos = strpos($contentType, ';')) !== false) {
+                // e.g. application/json; charset=UTF-8
+                $contentType = substr($contentType, 0, $pos);
+            }
+            if (isset($this->parsers[$contentType])) {
+                $this->_bodyParams = call_user_func($this->parsers[$contentType], $rawBody);
+            } elseif (isset($this->parsers['*'])) {
+                $this->_bodyParams = call_user_func($this->parsers['*'], $rawBody);
+            } elseif ($this->getMethod() === 'POST') {
+                $this->_bodyParams = $_POST;
+            } else {
+                mb_parse_str($rawBody, $this->_bodyParams);
+            }
+        }
+        if ($name === null) {
+            return $this->_bodyParams;
+        } else {
+            return isset($this->_bodyParams[$name]) ? $this->_bodyParams[$name] : $default;
+        }
+    }
+
+    public function get($name = null, $default = null)
+    {
+        $params = $this->getParams();
+        if ($name === null) {
+            return $params;
+        } else {
+            return isset($params[$name]) ? $params[$name] : $default;
+        }
+    }
+    private $_params;
+
+    public function getParams()
+    {
+        if ($this->_params === null) {
+            return $_GET;
+        }
+        return $this->_params;
+    }
+
+    public function setParams($value)
+    {
+        $this->_params = $value;
+    }
+    private $_cookies;
+
+    public function getCookie($name, $default = null)
+    {
+        if ($this->_cookies === null) {
+            $this->_cookies = [];
+            foreach ($_COOKIE as $name => $value) {
+                if (is_string($value) && $data = Helper::validateData($value, $this->cookieValidationKey) !== false) {
+                    $this->_cookies[$name] = $data;
+                }
+            }
+        }
+        return array_key_exists($this->_cookies, $name) ? $this->_cookies[$name] : $default;
+    }
+
+    /**
+     * Returns the method of the current request (e.g. GET, POST, HEAD, PUT, PATCH, DELETE).
+     * @return string request method, such as GET, POST, HEAD, PUT, PATCH, DELETE.
+     * The value returned is turned into upper case.
+     */
+    public function getMethod()
+    {
+        if (isset($_SERVER['HTTP_X_HTTP_METHOD_OVERRIDE'])) {
+            return strtoupper($_SERVER['HTTP_X_HTTP_METHOD_OVERRIDE']);
+        }
+
+        if (isset($_SERVER['REQUEST_METHOD'])) {
+            return strtoupper($_SERVER['REQUEST_METHOD']);
+        }
+
+        return 'GET';
+    }
+
+    /**
+     * Returns request content-type
+     * The Content-Type header field indicates the MIME type of the data
+     * contained in [[getRawBody()]] or, in the case of the HEAD method, the
+     * media type that would have been sent had the request been a GET.
+     * For the MIME-types the user expects in response, see [[acceptableContentTypes]].
+     * @return string request content-type. Null is returned if this information is not available.
+     * @link http://www.w3.org/Protocols/rfc2616/rfc2616-sec14.html#sec14.17
+     * HTTP 1.1 header field definitions
+     */
+    public function getContentType()
+    {
+        if (isset($_SERVER['CONTENT_TYPE'])) {
+            return $_SERVER['CONTENT_TYPE'];
+        } elseif (isset($_SERVER['HTTP_CONTENT_TYPE'])) {
+            //fix bug https://bugs.php.net/bug.php?id=66606
+            return $_SERVER['HTTP_CONTENT_TYPE'];
+        }
+
+        return null;
+    }
+    private $_hostInfo;
+
+    /**
+     * Returns the schema and host part of the current request URL.
+     * The returned URL does not have an ending slash.
+     * By default this is determined based on the user request information.
+     * You may explicitly specify it by setting the [[setHostInfo()|hostInfo]] property.
+     * @return string schema and hostname part (with port number if needed) of the request URL (e.g. `http://www.yiiframework.com`),
+     * null if can't be obtained from `$_SERVER` and wasn't set.
+     * @see setHostInfo()
+     */
+    public function getHostInfo()
+    {
+        if ($this->_hostInfo === null) {
+            $secure = $this->getIsSecureConnection();
+            $http = $secure ? 'https' : 'http';
+            if (isset($_SERVER['HTTP_HOST'])) {
+                $this->_hostInfo = $http . '://' . $_SERVER['HTTP_HOST'];
+            } elseif (isset($_SERVER['SERVER_NAME'])) {
+                $this->_hostInfo = $http . '://' . $_SERVER['SERVER_NAME'];
+                $port = $secure ? $this->getSecurePort() : $this->getPort();
+                if (($port !== 80 && !$secure) || ($port !== 443 && $secure)) {
+                    $this->_hostInfo .= ':' . $port;
+                }
+            }
+        }
+
+        return $this->_hostInfo;
+    }
+
+    /**
+     * Sets the schema and host part of the application URL.
+     * This setter is provided in case the schema and hostname cannot be determined
+     * on certain Web servers.
+     * @param string $value the schema and host part of the application URL. The trailing slashes will be removed.
+     */
+    public function setHostInfo($value)
+    {
+        $this->_hostInfo = rtrim($value, '/');
+    }
+    private $_baseUrl;
+
+    /**
+     * Returns the relative URL for the application.
+     * This is similar to [[scriptUrl]] except that it does not include the script file name,
+     * and the ending slashes are removed.
+     * @return string the relative URL for the application
+     * @see setScriptUrl()
+     */
+    public function getBaseUrl()
+    {
+        if ($this->_baseUrl === null) {
+            $this->_baseUrl = rtrim(dirname($this->getScriptUrl()), '\\/');
+        }
+
+        return $this->_baseUrl;
+    }
+
+    /**
+     * Sets the relative URL for the application.
+     * By default the URL is determined based on the entry script URL.
+     * This setter is provided in case you want to change this behavior.
+     * @param string $value the relative URL for the application
+     */
+    public function setBaseUrl($value)
+    {
+        $this->_baseUrl = $value;
+    }
+    private $_scriptUrl;
+
+    /**
+     * Returns the relative URL of the entry script.
+     * The implementation of this method referenced Zend_Controller_Request_Http in Zend Framework.
+     * @return string the relative URL of the entry script.
+     * @throws InvalidConfigException if unable to determine the entry script URL
+     */
+    public function getScriptUrl()
+    {
+        if ($this->_scriptUrl === null) {
+            $scriptFile = $this->getScriptFile();
+            $scriptName = basename($scriptFile);
+            if (isset($_SERVER['SCRIPT_NAME']) && basename($_SERVER['SCRIPT_NAME']) === $scriptName) {
+                $this->_scriptUrl = $_SERVER['SCRIPT_NAME'];
+            } elseif (isset($_SERVER['PHP_SELF']) && basename($_SERVER['PHP_SELF']) === $scriptName) {
+                $this->_scriptUrl = $_SERVER['PHP_SELF'];
+            } elseif (isset($_SERVER['ORIG_SCRIPT_NAME']) && basename($_SERVER['ORIG_SCRIPT_NAME']) === $scriptName) {
+                $this->_scriptUrl = $_SERVER['ORIG_SCRIPT_NAME'];
+            } elseif (isset($_SERVER['PHP_SELF']) && ($pos = strpos($_SERVER['PHP_SELF'], '/' . $scriptName)) !== false) {
+                $this->_scriptUrl = substr($_SERVER['SCRIPT_NAME'], 0, $pos) . '/' . $scriptName;
+            } elseif (!empty($_SERVER['DOCUMENT_ROOT']) && strpos($scriptFile, $_SERVER['DOCUMENT_ROOT']) === 0) {
+                $this->_scriptUrl = str_replace('\\', '/', str_replace($_SERVER['DOCUMENT_ROOT'], '', $scriptFile));
+            } else {
+                throw new InvalidConfigException('Unable to determine the entry script URL.');
+            }
+        }
+
+        return $this->_scriptUrl;
+    }
+
+    /**
+     * Sets the relative URL for the application entry script.
+     * This setter is provided in case the entry script URL cannot be determined
+     * on certain Web servers.
+     * @param string $value the relative URL for the application entry script.
+     */
+    public function setScriptUrl($value)
+    {
+        $this->_scriptUrl = '/' . trim($value, '/');
+    }
+    private $_scriptFile;
+
+    /**
+     * Returns the entry script file path.
+     * The default implementation will simply return `$_SERVER['SCRIPT_FILENAME']`.
+     * @return string the entry script file path
+     * @throws InvalidConfigException
+     */
+    public function getScriptFile()
+    {
+        if (isset($this->_scriptFile)) {
+            return $this->_scriptFile;
+        } elseif (isset($_SERVER['SCRIPT_FILENAME'])) {
+            return $_SERVER['SCRIPT_FILENAME'];
+        } else {
+            throw new InvalidConfigException('Unable to determine the entry script file path.');
+        }
+    }
+
+    /**
+     * Sets the entry script file path.
+     * The entry script file path normally can be obtained from `$_SERVER['SCRIPT_FILENAME']`.
+     * If your server configuration does not return the correct value, you may configure
+     * this property to make it right.
+     * @param string $value the entry script file path.
+     */
+    public function setScriptFile($value)
+    {
+        $this->_scriptFile = $value;
+    }
+    private $_pathInfo;
+
+    /**
+     * Returns the path info of the currently requested URL.
+     * A path info refers to the part that is after the entry script and before the question mark (query string).
+     * The starting and ending slashes are both removed.
+     * @return string part of the request URL that is after the entry script and before the question mark.
+     * Note, the returned path info is already URL-decoded.
+     * @throws InvalidConfigException if the path info cannot be determined due to unexpected server configuration
+     */
+    public function getPathInfo()
+    {
+        if ($this->_pathInfo === null) {
+            $this->_pathInfo = $this->resolvePathInfo();
+        }
+
+        return $this->_pathInfo;
+    }
+
+    /**
+     * Sets the path info of the current request.
+     * This method is mainly provided for testing purpose.
+     * @param string $value the path info of the current request
+     */
+    public function setPathInfo($value)
+    {
+        $this->_pathInfo = ltrim($value, '/');
+    }
+
+    /**
+     * Resolves the path info part of the currently requested URL.
+     * A path info refers to the part that is after the entry script and before the question mark (query string).
+     * The starting slashes are both removed (ending slashes will be kept).
+     * @return string part of the request URL that is after the entry script and before the question mark.
+     * Note, the returned path info is decoded.
+     * @throws InvalidConfigException if the path info cannot be determined due to unexpected server configuration
+     */
+    protected function resolvePathInfo()
+    {
+        $pathInfo = $this->getUrl();
+
+        if (($pos = strpos($pathInfo, '?')) !== false) {
+            $pathInfo = substr($pathInfo, 0, $pos);
+        }
+
+        $pathInfo = urldecode($pathInfo);
+
+        // try to encode in UTF8 if not so
+        // http://w3.org/International/questions/qa-forms-utf-8.html
+        if (!preg_match('%^(?:
+            [\x09\x0A\x0D\x20-\x7E]              # ASCII
+            | [\xC2-\xDF][\x80-\xBF]             # non-overlong 2-byte
+            | \xE0[\xA0-\xBF][\x80-\xBF]         # excluding overlongs
+            | [\xE1-\xEC\xEE\xEF][\x80-\xBF]{2}  # straight 3-byte
+            | \xED[\x80-\x9F][\x80-\xBF]         # excluding surrogates
+            | \xF0[\x90-\xBF][\x80-\xBF]{2}      # planes 1-3
+            | [\xF1-\xF3][\x80-\xBF]{3}          # planes 4-15
+            | \xF4[\x80-\x8F][\x80-\xBF]{2}      # plane 16
+            )*$%xs', $pathInfo)
+        ) {
+            $pathInfo = utf8_encode($pathInfo);
+        }
+
+        $scriptUrl = $this->getScriptUrl();
+        $baseUrl = $this->getBaseUrl();
+        if (strpos($pathInfo, $scriptUrl) === 0) {
+            $pathInfo = substr($pathInfo, strlen($scriptUrl));
+        } elseif ($baseUrl === '' || strpos($pathInfo, $baseUrl) === 0) {
+            $pathInfo = substr($pathInfo, strlen($baseUrl));
+        } elseif (isset($_SERVER['PHP_SELF']) && strpos($_SERVER['PHP_SELF'], $scriptUrl) === 0) {
+            $pathInfo = substr($_SERVER['PHP_SELF'], strlen($scriptUrl));
+        } else {
+            throw new InvalidConfigException('Unable to determine the path info of the current request.');
+        }
+
+        if (substr($pathInfo, 0, 1) === '/') {
+            $pathInfo = substr($pathInfo, 1);
+        }
+
+        return (string) $pathInfo;
+    }
+
+    /**
+     * Returns the currently requested absolute URL.
+     * This is a shortcut to the concatenation of [[hostInfo]] and [[url]].
+     * @return string the currently requested absolute URL.
+     */
+    public function getAbsoluteUrl()
+    {
+        return $this->getHostInfo() . $this->getUrl();
+    }
+    private $_url;
+
+    /**
+     * Returns the currently requested relative URL.
+     * This refers to the portion of the URL that is after the [[hostInfo]] part.
+     * It includes the [[queryString]] part if any.
+     * @return string the currently requested relative URL. Note that the URI returned is URL-encoded.
+     * @throws InvalidConfigException if the URL cannot be determined due to unusual server configuration
+     */
+    public function getUrl()
+    {
+        if ($this->_url === null) {
+            $this->_url = $this->resolveRequestUri();
+        }
+
+        return $this->_url;
+    }
+
+    /**
+     * Sets the currently requested relative URL.
+     * The URI must refer to the portion that is after [[hostInfo]].
+     * Note that the URI should be URL-encoded.
+     * @param string $value the request URI to be set
+     */
+    public function setUrl($value)
+    {
+        $this->_url = $value;
+    }
+
+    /**
+     * Resolves the request URI portion for the currently requested URL.
+     * This refers to the portion that is after the [[hostInfo]] part. It includes the [[queryString]] part if any.
+     * The implementation of this method referenced Zend_Controller_Request_Http in Zend Framework.
+     * @return string|boolean the request URI portion for the currently requested URL.
+     * Note that the URI returned is URL-encoded.
+     * @throws InvalidConfigException if the request URI cannot be determined due to unusual server configuration
+     */
+    protected function resolveRequestUri()
+    {
+        if (isset($_SERVER['HTTP_X_REWRITE_URL'])) { // IIS
+            $requestUri = $_SERVER['HTTP_X_REWRITE_URL'];
+        } elseif (isset($_SERVER['REQUEST_URI'])) {
+            $requestUri = $_SERVER['REQUEST_URI'];
+            if ($requestUri !== '' && $requestUri[0] !== '/') {
+                $requestUri = preg_replace('/^(http|https):\/\/[^\/]+/i', '', $requestUri);
+            }
+        } elseif (isset($_SERVER['ORIG_PATH_INFO'])) { // IIS 5.0 CGI
+            $requestUri = $_SERVER['ORIG_PATH_INFO'];
+            if (!empty($_SERVER['QUERY_STRING'])) {
+                $requestUri .= '?' . $_SERVER['QUERY_STRING'];
+            }
+        } else {
+            throw new InvalidConfigException('Unable to determine the request URI.');
+        }
+
+        return $requestUri;
+    }
+}
