+++ phpunit.xml.dist
@@ -1,0 +1,30 @@
+<?xml version="1.0" encoding="utf-8"?>
+<phpunit bootstrap="./tests/bootstrap.php"
+		colors="true"
+		convertErrorsToExceptions="true"
+		convertNoticesToExceptions="true"
+		convertWarningsToExceptions="true"
+		stopOnFailure="false">
+		<testsuites>
+			<testsuite name="Yii Test Suite">
+				<directory>./tests</directory>
+			</testsuite>
+		</testsuites>
+		<filter>
+			<blacklist>
+				<file>framework/helpers/Json.php</file>
+				<file>framework/helpers/StringHelper.php</file>
+				<file>framework/helpers/VarDumper.php</file>
+				<file>framework/helpers/Html.php</file>
+				<file>framework/helpers/Inflector.php</file>
+				<file>framework/helpers/FileHelper.php</file>
+				<file>framework/helpers/ArrayHelper.php</file>
+				<file>framework/helpers/Console.php</file>
+				<file>framework/i18n/GettextFile.php</file>
+				<file>framework/web/ResponseFormatterInterface.php</file>
+				<directory suffix="Exception.php">framework/base</directory>
+				<directory suffix=".php">framework/db/mssql</directory>
+				<directory suffix=".php">framework/bootstrap</directory>
+			</blacklist>
+		</filter>
+</phpunit>
